requirements.txt
Flask~=2.0.3
pyalsaaudio~=0.9.0
